requirements.txt
colorama
litellm
matplotlib
numpy
pandas
pydantic
python-dotenv
scikit-learn